+++ R/glm.R
@@ -59,6 +59,7 @@
 #' @examples
 #' library(recipes)
 #' library(dplyr)
+#' library(modeldata)
 #' 
 #' data(okc)
 #' 
@@ -78,7 +79,7 @@ step_lencode_glm <-
            skip = FALSE,
            id = rand_id("lencode_bayes")) {
     if (is.null(outcome))
-      stop("Please list a variable in `outcome`", call. = FALSE)
+      rlang::abort("Please list a variable in `outcome`")
     add_step(
       recipe,
       step_lencode_glm_new(