+++ tests/testthat/test_mixed.R
@@ -1,12 +1,9 @@
-library(embed)
-library(dplyr)
-library(testthat)
-
-source("make_example_data.R")
-
-###################################################################
-
 context("mixed model, classification")
+
+source(testthat::test_path("make_example_data.R"))
+source(testthat::test_path("test_helpers.R"))
+
+# ------------------------------------------------------------------------------
 
 test_that("factor encoded predictor", {
   class_test <- recipe(x2 ~ ., data = ex_dat) %>%